Moved Entry and Entry Utility
