+++ web/__init__.py
@@ -32,7 +32,7 @@ def create_app():
     from web.routes.character import character_bp
     app.register_blueprint(character_bp)
 
-    from web.entry.routes import entry_bp
+    from web.routes.entry import entry_bp
     app.register_blueprint(entry_bp)
 
     return app
